+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- [lighthouse][sdk] Fix URL generation
+
 ## Amarok 0.2.0-beta.6
 
 - [sdk] Fix URL